feat : ChromaDB 구축 자동화 함수 설정
- 데이터 파일 이름 변경
- Chroma DB 생성 코드(chroma_db.py) 함수화
- vector_store.py에 Chroma DB 생성 함수 추가
- Chroma DB 접근 디버깅 로그 작성

 #42
